fix : 프로젝트 코드 1차 병합으로 일부 코드 수정 ; issue description 참조
