+++ .idea/modules.xml
@@ -2,6 +2,7 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
+      <module fileurl="file://$PROJECT_DIR$/.idea/modules/com.pcproject.pcproject3.main.iml" filepath="$PROJECT_DIR$/.idea/modules/com.pcproject.pcproject3.main.iml" />
       <module fileurl="file://$PROJECT_DIR$/.idea/modules/pcproject3.iml" filepath="$PROJECT_DIR$/.idea/modules/pcproject3.iml" />
       <module fileurl="file://$PROJECT_DIR$/.idea/modules/pcproject3.main.iml" filepath="$PROJECT_DIR$/.idea/modules/pcproject3.main.iml" />
       <module fileurl="file://$PROJECT_DIR$/.idea/modules/pcproject3.test.iml" filepath="$PROJECT_DIR$/.idea/modules/pcproject3.test.iml" />
