--- src/main/java/com/pcproject/customer/entity/CustomerEntity.java
@@ -1,63 +1,0 @@
-package com.pcproject.customer.entity;
-
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.hibernate.annotations.SQLDelete;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import java.time.LocalDateTime;
-import java.util.ArrayList;
-
-@Getter
-@Entity
-@Table(name = "customers")
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@EntityListeners(AuditingEntityListener.class)
-@SQLDelete(sql = "UPDATE customers SET deleted_at = NOW(), status = 'WITHDRAWN' WHERE id = ?")
-public class CustomerEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private String name;
-
-    @Column(nullable = false, unique = true)
-    private String email;
-
-    private String phoneNumber;
-
-    @Enumerated(EnumType.STRING)
-    private CustomerStatus status = CustomerStatus.ACTIVE;
-
-    @CreatedDate
-    @Column(updatable = false)
-    private LocalDateTime createdAt;
-
-    private LocalDateTime deletedAt;
-
-    // orederEntity 임의 설정와 customerEntity 연결해야함. ********
-    @OneToMany(mappedBy = "customer", cascade = CascadeType.ALL, orphanRemoval = true)
-    private List<OrderEntity> orders = new ArrayList<>();
-
-    // 비즈니스 메서드 (정보 수정)
-    public void updateInfo(String name, String email, String phoneNumber) {
-        this.name = name;
-        this.email = email;
-        this.phoneNumber = phoneNumber;
-    }
-
-    public void changeStatus(CustomerStatus status) {
-        this.status = status;
-    }
-
-    public void softDelete() {
-        this.deletedAt = LocalDateTime.now();
-        this.status = CustomerStatus.WITHDRAWN; // 상태도 탈퇴로 변경
-    }
-}
-
