+++ src/main/resources/application.properties
@@ -1,20 +1,12 @@
 spring.application.name=pcproject3
 
-
-spring.datasource.url=jdbc:mysql://localhost:3306/jpa
-spring.datasource.username=root
-spring.datasource.password=12345678
-spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
-spring.jpa.hibernate.ddl-auto=create
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQLDialect
-spring.jpa.show-sql=true
-spring.jpa.properties.hibernate.format_sql=true
-spring.datasource.url=jdbc:mysql://localhost:3307/pc_commerce_master
+spring.datasource.url=jdbc:mysql://localhost:3306/JPA7?serverTimezone=Asia/Seoul&characterEncoding=UTF-8
 spring.datasource.username=root
 spring.datasource.password=12345678
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.jpa.hibernate.ddl-auto=update
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQLDialect
 spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.format_sql=true
-spring.sql.init.mode=always
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQLDialect
+
+spring.sql.init.mode=never
